--- containers/tefca-viewer/src/app/query/components/resultsViewAccordion/resultsTable.module.css
@@ -1,23 +1,0 @@
-.accordionWrapper > h3 {
-  font-size: 32px;
-  color: white !important;
-  background-color: #2e6276;
-
-  &:hover {
-    background-color: #224856;
-  }
-}
-
-.accordionWrapper > div {
-  margin: 0.1rem 0rem;
-  padding: 1rem 1.5rem;
-}
-
-.accordionHeading {
-  font-family: "Source Sans Pro Web", "Helvetica Neue", "Helvetica", "Roboto",
-    "Arial", "sans-serif" !important;
-  font-size: 1.46rem !important;
-  line-height: 1.1 !important;
-  padding: 0.75rem 0;
-  margin-bottom: 0.5rem !important;
-}
